+++ packages/onboarding/.env.example
@@ -1,0 +1,8 @@
+# Application port exposing the API
+APP_PORT=10884
+L1_PROVIDER_ENDPOINT=
+L2_PROVIDER_ENDPOINT=
+L2_FUNDER_PRIVATE_KEY=
+L1_PORTAL_ADDRESS=
+# The threshold of user funds on L2. Leave blank for default 0.01 ETH.
+DEPOSIT_FUNDING_THRESHOLD=
