+++ packages/onboarding/README.md
@@ -1,0 +1,25 @@
+# Specular Onboarding Services
+
+This package implements the user onboarding services for the bridge.
+
+Note: this services is stateless.
+
+## Install
+
+```bash
+pnpm install
+```
+
+## Setup
+
+```bash
+cp .env.example .env
+```
+
+Fill in the `.env` file with the correct values.
+
+## Run
+
+```bash
+pnpm start
+```
