+++ packages/onboarding/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@specularl2/legacy-bridge-relayer",
+  "name": "@specularl2/onbarding-services",
   "version": "0.0.0",
   "main": "src/index.ts",
   "license": "Apache-2.0",
@@ -21,7 +21,8 @@
     "fastify": "^4.15.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.5",
-    "typescript-collections": "^1.3.3"
+    "typescript-collections": "^1.3.3",
+    "@specularl2/sdk": "workspace:*"
   },
   "devDependencies": {
     "@ethersproject/providers": "^5.7.2",
