+++ packages/onboarding/src/onboarding/index.ts
@@ -1,7 +1,6 @@
 export * from "./config";
-export * from "./messanger";
+export * from "./messenger";
 export * from "./plugin";
 export * from "./service";
-export * from "./state";
 export * from "./types";
 export * from "./utils";
