--- packages/relayer/.env.example
@@ -1,9 +1,0 @@
-APP_PORT=10884
-L1_ORACLE_UPDATE_INTERVAL=10
-L1_PROVIDER_ENDPOINT=
-L2_PROVIDER_ENDPOINT=
-L2_RELAYER_PRIVATE_KEY=
-L2_FUNDER_PRIVATE_KEY=
-L1_ORACLE_ADDRESS=
-L1_PORTAL_ADDRESS=
-L2_PORTAL_ADDRESS=
