--- packages/relayer/README.md
@@ -1,24 +1,0 @@
-# Specular Legacy Bridge Relayer
-
-This package implements an external `L1Oracle` relayer and user onboarding services for the bridge.
-Once Magi supports `L1Oracle` update, the external relayer will be deprecated.
-
-## Install
-
-```bash
-pnpm install
-```
-
-## Setup
-
-```bash
-cp .env.example .env
-```
-
-Fill in the `.env` file with the correct values.
-
-## Run
-
-```bash
-pnpm start
-```
