+++ packages/sdk/package.json
@@ -9,6 +9,8 @@
     "dist/**/*"
   ],
   "scripts": {
+    "prettier": "prettier --write .",
+    "lint": "eslint . --fix",
     "build": "tsc",
     "gen:binding": "typechain --target ethers-v5 --out-dir src/types/contracts \"./node_modules/@specularl2/contracts/artifacts/src/**/*[!dbg].json\"",
     "preinstall": "npx only-allow pnpm",
@@ -18,6 +20,8 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@specularl2/contracts": "workspace:*",
+    "eslint": "^8.56.0",
+    "prettier": "^3.1.1",
     "typechain": "^8.1.1",
     "typescript": "^4.9.3"
   },
