+++ packages/sdk/src/constants.ts
@@ -1,0 +1,5 @@
+// Pre-deploy contract addresses
+export const l1OracleAddress =
+  "0x2A00000000000000000000000000000000000010" as const;
+export const l2PortalAddress =
+  "0x2A00000000000000000000000000000000000011" as const;
